backend/requirements.txt
fastapi
uvicorn
sqlalchemy
langchain
pytest
pytest-asyncio
pytest-cov
httpx
python-dotenv
